CASM-3506: Add IUF support
This change adds support for the IUF in the `load-ims-artifacts` image
in order to implement the ims-upload operation. It requires that the IUF
release distribution be mounted within the container, and that the `IUF`
and `IUF_RELEASE_PATH` environment variables be set. Other environment
variables must be set within the container as normal.

Existing `load-ims-artifacts` images should still be compatible with
this change. Additionally, `manifest.yaml` files from existing
`load-ims-artifacts` images can be used with the IUF ims-upload
operation, maintaining backwards compatibility.

This change also establishes a 2.0.0 release of the `ims-load-artifacts`
package for the IUF impelmentation.

Test Description:
Run IUF workflows using enumerated IUF manifests and existing content
directories. Ensure that content is uploaded as expected.